pyproject.toml
[tool.poetry]
name = "mdbd"
version = "0.1.0"
description = ""
authors = ["Stephen Nicholas Swatman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamdeck = "==0.9.1"
pydantic = "^1.10.7"
rich = "^13.3.3"
typer = "^0.7.0"
pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[tool.poetry.scripts]
mdbd = 'mdbd.cli:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
